+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## Release 0.18
 
-* #41  DataFrame.unique - produce DataFrame with non-repeating rows
+*  #41 DataFrame.unique - produce DataFrame with non-repeating rows
+* #184 DataFrame: rename "convertColumn(pos, Exp)" to "replaceColumn(pos, Exp)"
 * #186 Exp.in, Exp.notIn, Series.in, Series.notIn
 * #187 Exp.last(), Series.last()
 * #189 ExcelLoader: Duplicate empty column names